+++ app/src/main/kotlin/com/citysavior/android/navigation/CitySaviorNavHost.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
 import com.citysavior.android.presentation.auth.navigation.authGraph
-import com.citysavior.android.presentation.common.navigation.mainGraph
+import com.citysavior.android.presentation.main.navigation.mainGraph
 import com.citysavior.android.ui.CitySaviorAppState
 
 @Composable
