+++ presentation/src/main/kotlin/com/citysavior/android/presentation/main/home/navigation/HomeNavigation.kt
@@ -1,9 +1,9 @@
-package com.citysavior.android.presentation.home.navigation
+package com.citysavior.android.presentation.main.home.navigation
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
-import com.citysavior.android.presentation.home.base.HomeScreen
+import com.citysavior.android.presentation.main.home.HomeScreen
 
 const val HOME_GRAPH_ROUTE_PATTERN = "/home"
