+++ presentation/src/main/kotlin/com/citysavior/android/presentation/main/profile/component/AchievementItem.kt
@@ -1,7 +1,6 @@
-package com.citysavior.android.presentation.profile.base.component
+package com.citysavior.android.presentation.main.profile.component
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -12,7 +11,6 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -20,7 +18,7 @@ import com.citysavior.android.presentation.common.constant.Colors
 import com.citysavior.android.presentation.common.constant.Sizes
 import com.citysavior.android.presentation.common.constant.TextStyles
 import com.citysavior.android.presentation.common.utils.drawColoredShadow
-import com.citysavior.android.presentation.home.base.component.CustomProgress
+import com.citysavior.android.presentation.main.home.component.CustomProgress
 
 @Composable
 fun AchievementItem(
